// src/types/user/user.enum.ts
export enum RolesType {
  ADMIN = "admin",
  MEMBER = "member",
  LEADER = "leader"
}

export enum JoinStatus {
  PENDING = "pending",
  ACCEPTED = "accepted",
  REJECTED = "rejected",
  NULL = "null"
}

export enum RolesTeam {
  MEMBER = "member",
  LEADER = "leader"
}

export enum Gender {
  MALE = "male",
  FEMALE = "female"
}